HorseDAO and HorseDAOImpl touched up

+++ src/server/persistence/racer/RacerRepository.java
@@ -10,10 +10,10 @@ public interface RacerRepository
 {
   Horse create ( String name ) throws SQLException;
   Horse readByID ( int id ) throws SQLException;
-  Horse readBySpeed_min ( float speedMin ) throws SQLException;
-  Horse readBySpeed_max ( float speedMax ) throws SQLException;
+  Horse readBySpeed_min ( int speedMin ) throws SQLException;
+  Horse readBySpeed_max ( int speedMax ) throws SQLException;
   List<Horse> readByName ( String searchName ) throws SQLException;
-  List<Horse> readAll() throws SQLException;
+  List<Horse> readAll () throws SQLException;
   void updateName ( Horse horse ) throws SQLException;
   void updateSpeedMin ( Horse horse ) throws SQLException;
   void updateSpeedMax ( Horse horse ) throws SQLException;
